fix: load categories in edit form to match create form behavior
